@@ -13,11 +13,11 @@
     "axios": "^0.16.2",
     "browser-sync": "2.18.13",
     "chart.js": "2.7.1",
-    "cross-env": "^5.0.1",
+    "cross-env": "^5.2.0",
     "del": "3.0.0",
     "font-awesome": "4.7.0",
     "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^4.0.0",
+    "gulp-autoprefixer": "^4.1.0",
     "gulp-bower-src": "0.1.0",
     "gulp-concat": "2.6.1",
     "gulp-contains": "1.1.0",
@@ -31,16 +31,16 @@
     "gulp-uglify": "3.0.0",
     "jquery": "3.2.1",
     "laravel-mix": "^1.0",
-    "lodash": "^4.17.4",
+    "lodash": "^4.17.10",
     "pace-progress": "1.0.2",
     "popper.js": "1.12.6",
     "require-dir": "0.3.2",
     "run-sequence": "2.2.0",
     "simple-line-icons": "2.4.1",
-    "vue": "^2.1.10"
+    "vue": "^2.5.17"
   },
   "dependencies": {
     "bootstrap": "^4.0.0",
-    "vue-table-component": "^2.0.0-alpha.1"
+    "vue-table-component": "1.*"
   }
 }
