@@ -15,7 +15,9 @@ class UserController extends Controller
      */
     public function index()
     {
-        return view('admin.users.index');
+        $users = User::all();
+
+        return view('admin.users.index')->with(compact('users'));
     }
 
     /**
