@@ -15,7 +15,7 @@ class CreateCategoriesTable extends Migration
     {
         Schema::create('categories', function (Blueprint $table) {
             $table->increments('id');
-            $table->string('name').
+            $table->string('name');
             $table->string('type');
             $table->boolean('is_professional');
             $table->timestamps();
