@@ -8,16 +8,10 @@
 				<div class="card">
 					<div class="card-header"></div>
 					<div class="card-body">
-						<table class="table table-responsive-sm table-striped">
-							<thead>
-								<tr>
-									<td></td>
-									<td></td>
-									<td></td>
-									<td></td>
-								</tr>
-							</thead>
-						</table>
+						<table-component :data="{{ $users->toJson() }}" table-class="table table-striped table-responsive-sm" filter-input-class="mb-md-2 form-control col-md-3 float-right" :show-caption="false">
+							<table-column show="name" label="Name"></table-column>
+							<table-column show="email" label="Email"></table-column>
+						</table-component>
 					</div>
 				</div>
 			</div>
