@@ -21,7 +21,7 @@
   <!-- Main styles for this application -->
   <link href="{{ asset('css/style.css') }}" rel="stylesheet">
   <!-- Styles required by this views -->
-  <link rel="stylesheet" href="{{ asset('css/custom.css') }}">  
+  <link rel="stylesheet" href="{{ asset('css/custom.css') }}">
 </head>
 <!-- BODY options, add following classes to body to change options
 '.header-fixed' - Fixed Header
